organizing components and adding styles from quartz
