--- src/client/components/SignUp.jsx
@@ -1,1 +1,0 @@
-
