+++ src/client/styles/components/Sidenav.less
@@ -1,0 +1,31 @@
+.container{
+  background-color: #292E41;
+  //min-height: 100vh;
+}
+// #393D51
+
+.navItem{
+  background-color: inherit;
+  transition: background-color 1s;
+}
+.navItem:hover{
+  //TODO: Bug: Color draws rectangle over the corners
+  background-color: #292E41;
+}
+
+.card{
+  margin: 5px;
+  //TODO: Figure out how to disable
+}
+
+#navCard{
+  flex-grow: 1;
+}
+
+h6{
+  color: red;
+}
+
+CardTitle{
+  color: white;
+}
