+++ src/client/styles/pages/Dashboard.less
@@ -1,0 +1,15 @@
+//https://medium.com/@mollyseeley/a-simple-react-native-navigation-drawer-ba10fc203ad
+.Mining{
+  background-color: #DBDCDE;
+}
+
+.container{
+  background-color: #292E41;
+  position:absolute;
+  top:40px;
+  bottom:0px;
+  display: flex;
+  flex-direction: row;
+  align-items: stretch;
+  padding: 0;
+}
